requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=42.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyJWT>=2.8.0
python-multipart>=0.0.6
openai>=1.0.0
APScheduler>=3.10.0
